requirements.txt
djangorestframework~=3.15.2
djangorestframework-simplejwt
psycopg2-binary

Django~=5.1.7